app/src/main/java/com/abifarhan/moviecatalogue/ui/movie/MovieViewModel.kt
package com.abifarhan.moviecatalogue.ui.movie

import androidx.lifecycle.LiveData
import androidx.lifecycle.ViewModel
import androidx.paging.PagedList
import com.abifarhan.moviecatalogue.data.source.local.entity.MovieEntity
import com.abifarhan.moviecatalogue.data.MovieRepository
import com.abifarhan.moviecatalogue.vo.Resource

class MovieViewModel(private val movieRepository: MovieRepository) : ViewModel() {

    fun getMovie(): LiveData<Resource<PagedList<MovieEntity>>> =
        movieRepository.getAllMovies()
}
